# .well-known/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "SQL Playground Plugin",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.example.com/v1",
      "description": "Production server"
    },
    {
      "url": "https://api.staging.example.com/v1",
      "description": "Staging server"
    }
  ],
  "paths": {
    "/tables": {
      "get": {
        "summary": "Get Tables",
        "description": "Returns a list of tables in the database.",
        "operationId": "get_tables_tables_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{table_name}/columns": {
      "get": {
        "summary": "Get Columns",
        "description": "Returns a list of columns in a table.",
        "operationId": "get_columns_tables__table_name__columns_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Table Name",
              "type": "string"
            },
            "name": "table_name",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableColumnsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/execute_sql": {
      "post": {
        "summary": "Execute Sql",
        "description": "Executes a SQL query and returns the result.",
        "operationId": "execute_sql_execute_sql_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SQLQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SQLQueryResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "SQLQuery": {
        "title": "SQLQuery",
        "required": [
          "query"
        ],
        "type": "object",
        "properties": {
          "query": {
            "title": "Query",
            "type": "string"
          }
        }
      },
      "SQLQueryResult": {
        "title": "SQLQueryResult",
        "required": [
          "result"
        ],
        "type": "object",
        "properties": {
          "result": {
            "title": "Result",
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "TableColumn": {
        "title": "TableColumn",
        "required": [
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        }
      },
      "TableColumnsResponse": {
        "title": "TableColumnsResponse",
        "required": [
          "columns"
        ],
        "type": "object",
        "properties": {
          "columns": {
            "title": "Columns",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TableColumn"
            }
          }
        }
      },
      "TablesResponse": {
        "title": "TablesResponse",
        "required": [
          "tables"
        ],
        "type": "object",
        "properties": {
          "tables": {
            "title": "Tables",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}